[VS-7339] Program refactored to contain one startup class; API self health check and Dynamics Dataverse connection health check added
